Split Operands Validation
-> Might further migrate this with Operands Class
