+++ src/test/java/test_util/providers/Trim.java
@@ -1,0 +1,2 @@
+package test_util.providers;public class Trim {
+}
